fix the undefined behavior and the issue from big-endian, and extract half to another test
